refactor(analysis): extract complexity histogram helper
